refactor(mga): split micro_clip helpers into sibling module to clear file-size guard
CI file-size growth guard tripped on the previous commit (micro_clip_generator
went 618 → 787 LOC; orchestrator 735 → 745). Per per-app Tech Debt Policy
(mode: no-growth on flagged files), extract sibling instead of inlining:

  - new app/services/ingestion/micro_clip_helpers.py — _resolve_stand_ts_via_grid_classifier,
    _resolve_release_ts_via_throw_localizer, _cut_upload_persist_one_side. 273 LOC.
  - micro_clip_generator: 505 LOC (down from 618 base; was 787 inflated).
  - ingestion_orchestrator: 733 LOC (down from 735 base; collapsed multi-line
    ternary + trimmed comment).
  - tests: rebind helper-side patches to _HMOD (helpers module) — module-level
    objects like cut_clip / get_storage / classify_frames_for_lineup_decision
    now live in helpers, so patch targets follow.

679 backend tests still green.
