+++ apps/mygamingassistant/backend/app/services/ingestion/micro_clip_generator.py
@@ -1,73 +1,45 @@
-"""Stand + Aim micro-clip generator — 1s looped clips for the STAND and AIM
-panes of the 2×2 storyboard tile (PR6).
+"""Stand + Aim micro-clip generator — 1s looped clips for the storyboard.
 
-PR4 introduced the four-pane storyboard (STAND still | AIM still+anchor | THROW
-clip | LANDING text). PR5 replaced the LANDING text with a real landing clip.
-PR6 replaces the STAND and AIM stills with **1-second looping micro-clips**.
-
-**Anchor sources differ per side** (operator-tuned 2026-05-23 after AIM was
-shown to be random):
+PR6 introduced the 1-second STAND + AIM micro-clips replacing static stills.
+Anchor sources differ per side (operator-tuned 2026-05-23):
 
   - **STAND** anchors on the grid classifier's chosen ``stand`` frame
-    (``timestamps[stand_idx]``). The 9-frame grid samples evenly across the
-    chapter — coarse, but the STAND moment ("I am at the spot") is many seconds
-    long, so the grid reliably catches it.
-
-  - **AIM** anchors on **release_ts − _AIM_PRE_RELEASE_SECONDS** (0.8s before
-    the throw-localizer's release frame). The aim moment ("crosshair locked on
-    the alignment marker, immediately before release") is sub-second; the
-    9-frame grid is far too sparse to catch it reliably (~2-3s spacing on a
-    typical chapter), so Claude was picking a random distant frame for AIM.
-    The throw-timing classifier's dense pass (0.5s spacing around release)
-    already sees the locked-aim moment — its first 2 seconds before
-    release_ts IS the aim moment. So AIM derives off release_ts, not the
-    grid. The 1.0s AIM clip spans [release − 0.8s, release + 0.2s] — locked-
-    aim plus the first 0.2s of throw initiation.
+    (``timestamps[stand_idx]``). The 9-frame grid reliably catches the
+    "I am at the spot" window (many seconds long).
+  - **AIM** anchors on ``release_ts − _AIM_PRE_RELEASE_SECONDS`` (0.8s
+    before the throw-localizer's release frame). The aim moment is
+    sub-second; the grid is far too sparse to catch it reliably — Claude
+    was picking random distant frames. The throw-timing dense pass
+    (0.5s spacing around release) already sees the locked-aim moment.
 
 Two entry paths share one contract:
 
-  1. **Ingest path** — orchestrator already ran the grid classifier (for
-     ``stand_idx``) AND the throw-localizer (for ``release_ts``). It passes
-     ``precomputed_stand_ts`` (= ``timestamps[stand_idx]``) and
-     ``precomputed_release_ts`` (= the dense-pass release frame timestamp).
-     Generator skips both classifier calls entirely. Cost: zero extra Claude
-     spend; two ffmpeg cuts + two MinIO uploads per chapter.
+  1. **Ingest path** — orchestrator passes ``precomputed_stand_ts`` (from
+     its grid run) AND ``precomputed_release_ts`` (from the throw-localizer
+     it already ran for the THROW clip). Generator skips both Claude calls.
+  2. **Backfill path** — both precomputed values are None; generator runs
+     the grid classifier (stand) AND ``localize_throw_with_refinement``
+     (release → aim) itself. Cost: 1 grid + 1-2 throw-timing calls.
 
-  2. **Backfill path** — standalone CLI. Both precomputed values are None;
-     generator runs the grid classifier itself to recover ``stand_ts`` AND
-     calls ``localize_throw_with_refinement`` to recover ``release_ts``.
-     Cost: 1 grid call + 1-2 throw-timing calls per lineup. The two sides
-     remain independent: a stand-side failure NEVER rolls back the
-     already-committed aim clip, and vice versa.
-
-The AIM still + the persisted ``aim_anchor_x/y`` overlay coords are NOT
-recut here — they continue to derive from the grid classifier's aim frame.
-This is intentional Phase-1 scope: the AIM clip (what the operator actually
-sees in the default clip-mode view) is fixed; the brief poster flash and
-the rarely-used still-mode fallback are accepted-imperfections, listed in
-TECH_DEBT.md for a follow-up that re-extracts a single PNG at AIM_TS too.
-See ``project_mga_aim_anchor_center_screen_invariant.md`` in auto-memory.
-
-Cut + encode the muted MP4 via :func:`cut_clip` (re-uses the PR2 ffmpeg
-wrapper — same encode contract, same ``+faststart``), upload to MinIO under
-a deterministic key, persist the bare key via the matching repo setter.
-
-Idempotent: keys are ``pending/{video_id}/{int(chapter_start)}-stand-micro.mp4``
-and ``pending/{video_id}/{int(chapter_start)}-aim-micro.mp4``. Re-running
-overwrites the same objects instead of orphaning new ones. The DB writes are
-**two separate one-column commits** (``set_stand_clip_url`` +
-``set_aim_clip_url``) so a stand-side failure NEVER rolls back the already-
-committed aim clip, and vice versa. Mirrors PR2 / PR5 exactly.
+Per-side independence: a stand-side failure NEVER rolls back the already-
+committed aim side, and vice versa. The AIM still + persisted aim_anchor
+coords are NOT recut — phase-1 scope is the clip only; the brief poster
+flash + rarely-used still-mode fallback are accepted-imperfections. See
+``project_mga_aim_anchor_center_screen_invariant.md`` in auto-memory.
 
 Failure handling (rules/no-bandaid-solutions.md +
-rules/check-third-party-error-codes.md): yt-dlp / ffmpeg / Claude failures
-are captured with structured codes, logged at WARNING, and returned in
-``stand_error_codes`` / ``aim_error_codes``. ``*_clip_url`` is left NULL and
-the matching pane renders its still fallback. Nothing silent-fails.
+rules/check-third-party-error-codes.md): every yt-dlp / ffmpeg / Claude
+failure is captured with structured codes, logged at WARNING, and returned
+in the per-side error_codes; the matching ``*_clip_url`` is left NULL and
+the pane renders its still fallback. Nothing silent-fails.
+
+Sibling helpers (``_resolve_stand_ts_via_grid_classifier``,
+``_resolve_release_ts_via_throw_localizer``, ``_cut_upload_persist_one_side``)
+live in :mod:`micro_clip_helpers` to keep this file under the file-size
+growth guard (per per-app Tech Debt Policy).
 """
 from __future__ import annotations
 
-import asyncio
 import logging
 from dataclasses import dataclass, field
 from pathlib import Path
@@ -76,22 +48,15 @@ from typing import Optional
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from app.core.config import settings
-from app.core.storage import get_storage
 from app.models.game.lineup import Lineup
 from app.repositories.game import lineup_repo
-from app.services.classification.classifier_service import (
-    classify_frames_for_lineup_decision,
-)
 from app.services.ingestion.frame_extractor import (
-    ClipCutError,
-    FrameExtractionError,
-    cut_clip,
-    extract_frames,
-    grid_timestamps,
     wide_source_bounds,
 )
-from app.services.ingestion.throw_localizer import (
-    localize_throw_with_refinement,
+from app.services.ingestion.micro_clip_helpers import (
+    _cut_upload_persist_one_side,
+    _resolve_release_ts_via_throw_localizer,
+    _resolve_stand_ts_via_grid_classifier,
 )
 from app.services.ingestion.youtube_fetcher import (
     VideoDownloadError,
@@ -100,28 +65,17 @@ from app.services.ingestion.youtube_fetcher import (
 
 logger = logging.getLogger(__name__)
 
-# Frozen design-contract constants.
-# Per-pane micro-clip durations. STAND is 2.0s (operator-tuned: 1.0s was
-# cutting off mid-stance, before the player begins the throw motion). AIM
-# stays at 1.0s — the AIM pane's whole point is the locked-aim moment;
-# longer would bleed deeper into the throw animation (which is what the
-# THROW pane already shows).
+# Per-pane micro-clip durations. STAND is 2.0s (operator-tuned: 1.0s cut
+# off mid-stance). AIM stays at 1.0s — longer bleeds into the THROW pane's
+# window.
 _STAND_MICRO_CLIP_SECONDS = 2.0
 _AIM_MICRO_CLIP_SECONDS = 1.0
-# Seconds BEFORE release_ts the AIM clip starts at. Result: the 1.0s AIM
-# clip spans [release − 0.8s, release + 0.2s] — locked-aim moment with a
-# tiny peek of throw initiation as a natural end-cue. Smaller would risk
-# dropping the lock-on frame on edge cases (slight release-frame error
-# from the dense pass); larger would step on the THROW pane's window.
+# Seconds BEFORE release_ts the AIM clip starts at. The 1.0s AIM clip
+# spans [release − 0.8s, release + 0.2s] — locked-aim with a tiny throw-
+# initiation peek as a natural end-cue.
 _AIM_PRE_RELEASE_SECONDS = 0.8
-# Minimum acceptable clamped duration. A near-end timestamp may clip short;
-# under this threshold we skip rather than ship a useless ~200ms sliver.
+# Minimum acceptable clamped duration. Below: skip rather than ship a sliver.
 _MIN_CLIP_SECONDS = 0.5
-# Grid sample count + edge padding — must match the ingest orchestrator's
-# ingestion_orchestrator._GRID_FRAME_COUNT / _GRID_EDGE_PADDING_SECONDS so
-# the backfill recovers the SAME timestamps the ingest pass extracted.
-_GRID_FRAME_COUNT = 9
-_GRID_EDGE_PADDING_SECONDS = 1.0
 
 
 @dataclass
@@ -222,139 +176,6 @@ def _micro_clip_seconds_for_side(side: str) -> float:
     if side == "aim":
         return _AIM_MICRO_CLIP_SECONDS
     raise ValueError(f"unknown micro-clip side: {side!r}")
-
-
-async def _resolve_stand_ts_via_grid_classifier(
-    db: AsyncSession,
-    lineup: Lineup,
-    video_path: Path,
-    chapter_start: float,
-    chapter_end: float,
-) -> tuple[Optional[float], list[str], str]:
-    """Backfill helper: re-run the grid classifier to recover stand_ts.
-
-    Returns ``(stand_ts, error_codes, reasoning)``. The grid classifier also
-    returns an ``aim_idx`` but it is intentionally discarded — AIM_TS is now
-    derived from release_ts (see module docstring); the grid pass is too
-    sparse to localise the sub-second aim moment. STAND is unaffected
-    because the standing-at-spot window is many seconds long and the
-    9-frame grid reliably catches it.
-    """
-    timestamps = grid_timestamps(
-        float(chapter_start),
-        float(chapter_end),
-        _GRID_FRAME_COUNT,
-        edge_padding_seconds=_GRID_EDGE_PADDING_SECONDS,
-    )
-    if not timestamps:
-        return None, ["empty_grid_window"], "chapter window produced no grid frames"
-
-    try:
-        frames = await extract_frames(video_path, timestamps)
-    except FrameExtractionError as exc:
-        logger.warning(
-            "micro_clip_generator: grid frame extraction failed: "
-            "lineup=%s returncode=%s stderr=%s",
-            lineup.id, exc.returncode, exc.stderr[:300],
-        )
-        return None, [f"frame_extract:rc={exc.returncode}"], str(exc)
-
-    if not frames:
-        return None, ["grid_frames_empty"], "ffmpeg returned no frames"
-
-    try:
-        result = await classify_frames_for_lineup_decision(
-            db,
-            frames=frames,
-            chapter_title=lineup.chapter_title or "",
-            attribution_author=lineup.attribution_author,
-            game_hint=None,
-        )
-    except Exception as exc:
-        logger.warning(
-            "micro_clip_generator: grid classifier raised: lineup=%s error=%s",
-            lineup.id, str(exc),
-        )
-        return None, [f"classifier_raised:{type(exc).__name__}"], str(exc)
-
-    if not result.success:
-        logger.warning(
-            "micro_clip_generator: grid classifier call failed: lineup=%s "
-            "error_codes=%s",
-            lineup.id, result.error_codes,
-        )
-        return None, list(result.error_codes), "grid classifier reported failure"
-
-    if not result.is_lineup:
-        # The backfill candidate is an accepted lineup, so the classifier
-        # judging it "not a lineup" here would be a regression — keep the
-        # signal but skip cleanly rather than fabricate a timestamp.
-        return None, [], "backfill grid says not_a_lineup (skip)"
-
-    # 1-based → 0-based, clamped to the grid range. Falls back to first
-    # index when the model omitted it — same shape as the orchestrator's
-    # classifier-disabled fallback.
-    stand_idx = (result.best_stand_index or 1) - 1
-    stand_idx = max(0, min(stand_idx, len(timestamps) - 1))
-    return timestamps[stand_idx], [], result.reasoning or ""
-
-
-async def _resolve_release_ts_via_throw_localizer(
-    lineup: Lineup,
-    video_path: Path,
-    chapter_start: float,
-    chapter_end: float,
-) -> tuple[Optional[float], list[str], str]:
-    """Backfill helper: run the throw-timing classifier to recover release_ts.
-
-    Returns ``(release_ts, error_codes, reasoning)``. AIM_TS is then
-    ``release_ts - _AIM_PRE_RELEASE_SECONDS`` (see module docstring on why
-    AIM derives from release, not the grid). When the localizer judges the
-    chapter not-a-throw or returns no release frame, we surface that as an
-    empty error_codes + structured reasoning (skip cleanly, not fail) — the
-    AIM clip is best-effort and a non-throw lineup is a real signal, not an
-    operational fault.
-    """
-    try:
-        refined = await localize_throw_with_refinement(
-            video_path,
-            chapter_start=float(chapter_start),
-            chapter_end=float(chapter_end),
-            chapter_title=lineup.chapter_title or "",
-            utility_hint=None,
-        )
-    except FrameExtractionError as exc:
-        # Coarse frame-extract failure propagates per throw_localizer's
-        # contract — capture as a structured code so the side-tally surfaces
-        # it instead of crashing.
-        logger.warning(
-            "micro_clip_generator: throw localizer extract failed: "
-            "lineup=%s returncode=%s stderr=%s",
-            lineup.id, exc.returncode, exc.stderr[:300],
-        )
-        return None, [f"throw_localizer_extract:rc={exc.returncode}"], str(exc)
-    except Exception as exc:  # defensive
-        logger.warning(
-            "micro_clip_generator: throw localizer raised: lineup=%s error=%s",
-            lineup.id, str(exc),
-        )
-        return None, [f"throw_localizer_raised:{type(exc).__name__}"], str(exc)
-
-    timing = refined.timing
-    if not timing.success:
-        logger.warning(
-            "micro_clip_generator: throw localizer call failed: lineup=%s "
-            "error_codes=%s",
-            lineup.id, timing.error_codes,
-        )
-        return None, list(timing.error_codes), "throw localizer reported failure"
-
-    if not timing.is_lineup_throw or timing.release_index is None:
-        # Not a throw OR no usable release frame — clean skip, not a fault.
-        return None, [], "throw localizer found no usable release frame"
-
-    release_ts = refined.frame_timestamps[timing.release_index - 1]
-    return release_ts, [], timing.reasoning or ""
 
 
 async def generate_micro_clips_for_lineup(
@@ -628,12 +449,14 @@ async def generate_micro_clips_for_lineup(
             stand_outcome = await _cut_upload_persist_one_side(
                 db, lineup, video_id, local_video,
                 anchor_ts=stand_ts,
+                clip_seconds=_STAND_MICRO_CLIP_SECONDS,
                 chapter_start=chapter_start,
                 chapter_end=chapter_end,
                 side="stand",
                 key_fn=pending_stand_clip_key,
                 persist_fn=lineup_repo.set_stand_clip_url,
                 wider_source_start_s=wider_source_start_s,
+                min_clip_seconds=_MIN_CLIP_SECONDS,
             )
         if aim_ts is None:
             aim_outcome = {
@@ -645,12 +468,14 @@ async def generate_micro_clips_for_lineup(
             aim_outcome = await _cut_upload_persist_one_side(
                 db, lineup, video_id, local_video,
                 anchor_ts=aim_ts,
+                clip_seconds=_AIM_MICRO_CLIP_SECONDS,
                 chapter_start=chapter_start,
                 chapter_end=chapter_end,
                 side="aim",
                 key_fn=pending_aim_clip_key,
                 persist_fn=lineup_repo.set_aim_clip_url,
                 wider_source_start_s=wider_source_start_s,
+                min_clip_seconds=_MIN_CLIP_SECONDS,
             )
 
         return MicroClipGenerationResult(
@@ -678,110 +503,3 @@ async def generate_micro_clips_for_lineup(
                 )
 
 
-async def _cut_upload_persist_one_side(
-    db: AsyncSession,
-    lineup: Lineup,
-    video_id: str,
-    local_video: Path,
-    *,
-    anchor_ts: float,
-    chapter_start: float,
-    chapter_end: float,
-    side: str,
-    key_fn,
-    persist_fn,
-    wider_source_start_s: float | None = None,
-) -> dict:
-    """Cut + upload + persist exactly ONE side (stand or aim).
-
-    Returns a flat dict consumed by ``generate_micro_clips_for_lineup`` —
-    not a dataclass because the caller flattens the two sides into the
-    composite ``MicroClipGenerationResult``. The same shape is used for
-    both sides so a future third pane (unlikely but possible) can be added
-    without touching this function.
-
-    *wider_source_start_s* is the seconds-into-source-video start of the
-    SHARED wider clip (``clip_url_original``) the operator's shift-window
-    slider will be positioned against. When set, this function computes
-    ``offset_s = clip_start - wider_source_start_s`` and persists it via the
-    setter's ``offset_s=`` kwarg so the shift overlay opens at the right
-    initial thumb position. When None (no wider source for this lineup),
-    the offset is left untouched in the DB and the shift overlay opens at 0.
-    """
-    bounds = _compute_micro_bounds(
-        anchor_ts,
-        chapter_start,
-        chapter_end,
-        clip_seconds=_micro_clip_seconds_for_side(side),
-    )
-    if bounds is None:
-        return {
-            "status": "skipped",
-            "skip_reason": "chapter_too_short_for_microclip",
-        }
-    clip_start, clip_duration = bounds
-
-    try:
-        clip_bytes = await cut_clip(local_video, clip_start, clip_duration)
-    except ClipCutError as exc:
-        logger.warning(
-            "micro_clip_generator: %s clip cut failed: lineup=%s video_id=%s "
-            "start=%.2f dur=%.2f returncode=%s stderr=%s",
-            side, lineup.id, video_id, clip_start, clip_duration,
-            exc.returncode, exc.stderr[:300],
-        )
-        return {
-            "status": "failed",
-            "error_codes": [f"clip_cut:rc={exc.returncode}"],
-        }
-
-    clip_key = key_fn(video_id, chapter_start)
-    try:
-        storage = get_storage()
-        loop = asyncio.get_running_loop()
-        await loop.run_in_executor(
-            None, storage.upload_file, clip_key, clip_bytes, "video/mp4"
-        )
-    except Exception as exc:
-        logger.warning(
-            "micro_clip_generator: %s clip upload failed: lineup=%s key=%s "
-            "error=%s",
-            side, lineup.id, clip_key, str(exc),
-        )
-        return {
-            "status": "failed",
-            "error_codes": ["clip_upload_failed"],
-        }
-
-    # Compute the in-source offset (only when a shared wider source exists for
-    # this lineup). The setter writes ``*_clip_offset_s`` so the PR2 shift
-    # overlay opens its slider at the position the ingest pipeline cut from.
-    if wider_source_start_s is not None:
-        offset_s: float | None = clip_start - wider_source_start_s
-    else:
-        offset_s = None
-
-    try:
-        await persist_fn(db, lineup, clip_key, offset_s=offset_s)
-    except Exception as exc:
-        # Object uploaded but column did not commit. The key is deterministic
-        # so a later backfill recomputes the same key and overwrites the same
-        # object — no orphan, safe to retry.
-        logger.warning(
-            "micro_clip_generator: %s_clip_url persist failed (object "
-            "uploaded, column not committed; backfill is idempotent): "
-            "lineup=%s key=%s error=%s",
-            side, lineup.id, clip_key, str(exc),
-        )
-        return {
-            "status": "failed",
-            "error_codes": [f"{side}_clip_url_persist_failed"],
-        }
-
-    logger.info(
-        "micro_clip_generator: %s clip generated: lineup=%s video_id=%s "
-        "key=%s anchor_ts=%.2f clip=[%.2f,+%.2fs]",
-        side, lineup.id, video_id, clip_key, anchor_ts,
-        clip_start, clip_duration,
-    )
-    return {"status": "generated", "clip_key": clip_key}
